# api/openapi.yaml
openapi: "3.1.0"
info:
  title: NovaFabric Multi-User REST API
  version: "0.7.0"
  description: |
    Multi-user REST API for the NovaFabric AI Asset Registry.
    URL prefix: /v0/
    Authentication: Bearer token via Authorization header (OIDC) or open when OIDC disabled.
    All list endpoints use cursor-based pagination (limit + next_cursor).
    All error responses use the standard error envelope.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /v0
    description: NovaFabric server (versioned)

# --------------------------------------------------------------------------- #
# Reusable components
# --------------------------------------------------------------------------- #
components:
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: Asset 'my-agent@v1.0' not found.
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"

    # -------- pagination --------
    PaginationMeta:
      type: object
      required: [total]
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor for the next page; null when no more pages.
        total:
          type: integer
          description: Total count of matching items (without pagination applied).

    # -------- assets --------
...
    LineageNode:
      type: object
      required: [node_id, kind, ref]
      properties:
        node_id:
          type: string
        kind:
          type: string
          enum: [run, asset, artifact, model, tool]
        ref:
          type: string
        payload:
          type: object
          additionalProperties: true
    LineageNodeListResponse:
      allOf:
        - $ref: "#/components/schemas/PaginationMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/LineageNode"
    LineageQueryResponse:
      type: object
      required: [ref, depth, nodes]
      properties:
        ref:
          type: string
        kind:
          type: string
          nullable: true
        depth:
          type: integer
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/LineageNode"
    ReplayChainResponse:
      type: object
      required: [run_id, chain]
      properties:
        run_id:
          type: string
        chain:
          type: array
          items:
            $ref: "#/components/schemas/LineageNode"

    # -------- replays --------
    ReplayMode:
      type: string
      enum: [forensic, mocked, shadow, dry_run]
    ReplayScheduleRequest:
      type: object
      required: [run_id, mode]
      properties:
        run_id:
          type: string
          description: The capsule run_id to replay.
        mode:
          $ref: "#/components/schemas/ReplayMode"
    ReplayStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
    ReplaySummary:
      type: object
      required: [replay_id, run_id, mode, status, created_at]
      properties:
        replay_id:
          type: string
        run_id:
          type: string
        mode:
          $ref: "#/components/schemas/ReplayMode"
        status:
          $ref: "#/components/schemas/ReplayStatus"
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true

    # -------- evidence --------
...
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
      description: Maximum number of items to return (default 50, max 500).
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
      required: false
      description: Opaque pagination cursor returned from a previous list call.
  responses:
    BadRequest:
      description: Invalid request parameters or body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"

# --------------------------------------------------------------------------- #
# Paths
# --------------------------------------------------------------------------- #
paths:
...
  /lineage/nodes:
    get:
      operationId: listLineageNodes
      summary: List lineage nodes
      tags: [lineage]
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/CursorParam"
        - name: kind
          in: query
          required: false
          schema:
            type: string
            enum: [run, asset, artifact, model, tool]
      responses:
        "200":
          description: Paginated list of lineage nodes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineageNodeListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /lineage/blast-radius:
    get:
      operationId: lineageBlastRadius
      summary: Compute blast radius (downstream dependents) from a node reference
      tags: [lineage]
      parameters:
        - name: ref
          in: query
          required: true
          schema:
            type: string
          description: Node reference string (run_id, asset ref, artifact ref).
        - name: kind
          in: query
          required: false
          schema:
            type: string
          description: Node kind filter (e.g. run, asset).
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: Blast-radius nodes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineageQueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /lineage/provenance:
    get:
      operationId: lineageProvenance
      summary: Compute provenance (upstream ancestors) from a node reference
      tags: [lineage]
      parameters:
        - name: ref
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: false
          schema:
            type: string
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            default: 5
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: Provenance nodes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineageQueryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /lineage/replay-chain:
    get:
      operationId: lineageReplayChain
      summary: Get the replay chain for a run
      tags: [lineage]
      parameters:
        - name: run_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay chain nodes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayChainResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ======================== REPLAYS ========================
  /replays:
    post:
      operationId: scheduleReplay
      summary: Schedule a replay job
      tags: [replays]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplayScheduleRequest"
      responses:
        "202":
          description: Replay job accepted; check status via GET /replays/{replay_id}.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaySummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /replays/{replay_id}:
    get:
      operationId: getReplay
      summary: Get replay job status
      tags: [replays]
      parameters:
        - name: replay_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaySummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: cancelReplay
      summary: Cancel a pending or running replay job
      tags: [replays]
      parameters:
        - name: replay_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Replay cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaySummary"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Replay already completed or cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /replays/{replay_id}/events:
    get:
      operationId: replayEvents
      summary: Stream replay progress as Server-Sent Events
      tags: [replays]
      parameters:
        - name: replay_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: |
            SSE stream. Each event is a JSON object:
            `{"event": "progress|completed|error", "data": {...}}`.
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/NotFound"

  # ======================== EVIDENCE ========================
...